feat publish element

+++ partridge-mng-service/src/main/java/cn/nyaaar/partridgemngservice/controller/UploadController.java
@@ -37,7 +37,7 @@ public class UploadController {
     @Operation(summary = "检查文件", description = "检查文件分片是否全部上传完成")
     @PostMapping(value = "/check")
     @LogAnnotation
-    public R<CheckResp> check(@RequestBody Integer eleId) {
+    public R<CheckResp> check(@RequestBody Long eleId) {
         return new R<>(uploadService.check(eleId));
     }
