+++ partridge-mng-service/src/main/java/cn/nyaaar/partridgemngservice/service/EleFileService.java
@@ -64,5 +64,5 @@ public interface EleFileService extends IService<EleFile> {
      * @param id id
      * @return EleFile
      */
-    EleFile findById(Integer id);
+    EleFile findById(Long id);
 }
